back/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.1.0
langchain==0.1.0
langchain-google-genai==0.0.5
pinecone-client==3.0.0
google-generativeai==0.3.1 